requirements.txt
pyPS4Controller==1.2.5
pyserial==3.5
